backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
# psycopg2-binary>=2.9.9   # Uncomment when using PostgreSQL
pydantic>=2.7.1
pydantic-settings>=2.2.1
PyJWT>=2.8.0               # JWT tokens (python-jose alternative for Python 3.14+)
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
aiosqlite>=0.19.0          # SQLite async support for tests
